Extract worker session recovery
